+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Description: Flexiv Rizon4 (Xacro)
 - Description: Robotiq 2F-85 v4 (URDF) (thanks to @nickswalker)
+- Loaders: Allow overriding `XACRO_ARGS` via `load_robot_description` (thanks to @nickswalker)
 
 ## [1.23.0] - 2026-03-09